# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Adicara API
  version: 0.3.0
  description: Canonical REST contract for the Adicara platform.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
    description: Same-origin API
...
security: []
paths:
...
  /invitations/{id}/guests:
    parameters:
      - $ref: '#/components/parameters/InvitationID'
    get:
      operationId: listGuests
      summary: List guests for an owned invitation
      tags: [Invitations]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Guests including their private management fields and public tokens.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [guests]
                properties:
                  guests:
                    type: array
                    items:
                      $ref: '#/components/schemas/Guest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createGuest
      summary: Add a guest to an owned invitation
      tags: [Invitations]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestWrite'
      responses:
        '201':
          description: Guest created with a secure random public token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /invitations/{id}/guests/{guestId}:
    parameters:
      - $ref: '#/components/parameters/InvitationID'
      - $ref: '#/components/parameters/GuestID'
    patch:
      operationId: updateGuest
      summary: Replace editable fields for an owned invitation guest
      tags: [Invitations]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestWrite'
      responses:
        '200':
          description: Guest updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      operationId: deleteGuest
      summary: Delete a guest from an owned invitation
      tags: [Invitations]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Guest deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: adicara_session
      description: Opaque HttpOnly session token. Only its SHA-256 hash is stored server-side.
  parameters:
    InvitationID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GuestID:
      name: guestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Per-session CSRF token returned after registration or login and mirrored in the non-HttpOnly adicara_csrf cookie.
      schema:
        type: string
        minLength: 32
  responses:
...
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: CSRF validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The owner-scoped resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Valid JSON failed domain validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    GuestWrite:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        group:
          type: string
          maxLength: 120
        phone:
          type: string
          maxLength: 40
        notes:
          type: string
          maxLength: 500
    Guest:
      allOf:
        - $ref: '#/components/schemas/GuestWrite'
        - type: object
          additionalProperties: false
          required: [id, public_token, rsvp_status, created_at, updated_at]
          properties:
            id:
              type: string
              format: uuid
            public_token:
              type: string
              minLength: 43
              maxLength: 43
            rsvp_status:
              type: string
              enum: [hadir, tidak_hadir, belum_menjawab]
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string